Grid Coordinate System + Mouse Click Detection
-Level is split into multiple grids.

-Left Mouse Click will detect the closest grid coordinate.

-Grid Coordinates returned are adjusted and turned from screen-space coordinates to world-space coordinates, taking into account the player position.

+++ PlatformerGameAE/Source/Structs.hpp
@@ -85,6 +85,6 @@ struct LavaSpout {
 };
 
 struct GridCoordinate {
-	f32 x, y;
+	s32 x, y;
 };
